lab3/requirements.txt
Django~=3.2
psycopg2~=2.9.1
pillow~=8.3
